Revert "implement data retention for boards (#2588)" (#2843)
This reverts commit 0489de8bd36b9b28a3f520f6b2c4c21dc4dc2f37.
