+++ server/services/config/config.go
@@ -50,8 +50,6 @@ type Configuration struct {
 	LocalModeSocketLocation  string            `json:"localModeSocketLocation" mapstructure:"localModeSocketLocation"`
 	EnablePublicSharedBoards bool              `json:"enablePublicSharedBoards" mapstructure:"enablePublicSharedBoards"`
 	FeatureFlags             map[string]string `json:"featureFlags" mapstructure:"featureFlags"`
-	EnableDataRetention      bool              `json:"enable_data_retention" mapstructure:"enable_data_retention"`
-	DataRetentionDays        int               `json:"data_retention_days" mapstructure:"data_retention_days"`
 
 	AuthMode string `json:"authMode" mapstructure:"authMode"`
 
@@ -97,8 +95,6 @@ func ReadConfigFile(configFilePath string) (*Configuration, error) {
 	viper.SetDefault("AuthMode", "native")
 	viper.SetDefault("NotifyFreqCardSeconds", 120)    // 2 minutes after last card edit
 	viper.SetDefault("NotifyFreqBoardSeconds", 86400) // 1 day after last card edit
-	viper.SetDefault("EnableDataRetention", false)
-	viper.SetDefault("DataRetentionDays", 365) // 1 year is default
 	viper.SetDefault("PrometheusAddress", "")
 
 	err := viper.ReadInConfig() // Find and read the config file
