+++ server/services/store/sqlstore/blocks.go
@@ -4,13 +4,9 @@ import (
 	"database/sql"
 	"encoding/json"
 	"fmt"
-	"strconv"
-	"strings"
-	"time"
 
 	"github.com/mattermost/focalboard/server/services/store"
 	"github.com/mattermost/focalboard/server/utils"
-	"github.com/pkg/errors"
 
 	sq "github.com/Masterminds/squirrel"
 	_ "github.com/lib/pq" // postgres driver
@@ -746,113 +742,6 @@ func (s *SQLStore) replaceBlockID(db sq.BaseRunner, currentID, newID, workspaceI
 	}
 
 	return nil
-}
-
-func (s *SQLStore) runDataRetention(db sq.BaseRunner, globalRetentionDate int64, batchSize int64) (int64, error) {
-	mlog.Info("Start Boards Data Retention",
-		mlog.String("Global Retention Date", time.Unix(globalRetentionDate/1000, 0).String()),
-		mlog.Int64("Raw Date", globalRetentionDate))
-	deleteTables := map[string]string{
-		"blocks":         "board_id",
-		"blocks_history": "board_id",
-		"boards":         "id",
-		"boards_history": "id",
-		"board_members":  "board_id",
-		"sharing":        "id",
-	}
-
-	subBuilder := s.getQueryBuilder(db).
-		Select("board_id, MAX(update_at) AS maxDate").
-		From(s.tablePrefix + "blocks").
-		GroupBy("board_id")
-
-	subQuery, _, _ := subBuilder.ToSql()
-
-	builder := s.getQueryBuilder(db).
-		Select("id").
-		From(s.tablePrefix + "boards").
-		LeftJoin("( " + subQuery + " ) As subquery ON (subquery.board_id = id)").
-		Where(sq.Lt{"maxDate": globalRetentionDate}).
-		Where(sq.NotEq{"team_id": "0"}).
-		Where(sq.Eq{"is_template": false})
-
-	rows, err := builder.Query()
-	if err != nil {
-		s.logger.Error(`dataRetention subquery ERROR`, mlog.Err(err))
-		return 0, err
-	}
-	defer s.CloseRows(rows)
-	deleteIds, err := idsFromRows(rows)
-	if err != nil {
-		return 0, err
-	}
-
-	totalAffected := 0
-	if len(deleteIds) > 0 {
-		mlog.Debug("Data Retention DeleteIDs " + strings.Join(deleteIds, ", "))
-		for table, field := range deleteTables {
-			affected, err := s.genericRetentionPoliciesDeletion(db, table, field, deleteIds, batchSize)
-			if err != nil {
-				return int64(totalAffected), err
-			}
-			totalAffected += int(affected)
-		}
-	}
-	mlog.Info("Complete Boards Data Retention", mlog.Int("total deletion ids", len(deleteIds)), mlog.Int("TotalAffected", totalAffected))
-	return int64(totalAffected), nil
-}
-
-func idsFromRows(rows *sql.Rows) ([]string, error) {
-	deleteIds := []string{}
-	for rows.Next() {
-		var boardID string
-		err := rows.Scan(
-			&boardID,
-		)
-		if err != nil {
-			return nil, err
-		}
-		deleteIds = append(deleteIds, boardID)
-	}
-	return deleteIds, nil
-}
-
-// genericRetentionPoliciesDeletion actually executes the DELETE query using a sq.SelectBuilder
-// which selects the rows to delete.
-func (s *SQLStore) genericRetentionPoliciesDeletion(
-	db sq.BaseRunner,
-	table string,
-	deleteColumn string,
-	deleteIds []string,
-	batchSize int64,
-) (int64, error) {
-	whereClause := deleteColumn + ` IN ('` + strings.Join(deleteIds, `','`) + `')`
-	deleteQuery := s.getQueryBuilder(db).
-		Delete(s.tablePrefix + table).
-		Where(whereClause)
-	s1, _, _ := deleteQuery.ToSql()
-	mlog.Debug(s1)
-
-	var totalRowsAffected int64
-	var batchRowsAffected int64
-	for {
-		result, err := deleteQuery.Exec()
-		if err != nil {
-			return 0, errors.Wrap(err, "failed to delete "+table)
-		}
-
-		batchRowsAffected, err = result.RowsAffected()
-		if err != nil {
-			return 0, errors.Wrap(err, "failed to get rows affected for "+table)
-		}
-		totalRowsAffected += batchRowsAffected
-		if batchRowsAffected != batchSize {
-			break
-		}
-	}
-
-	mlog.Debug("Rows Affected" + strconv.FormatInt(totalRowsAffected, 10))
-	return totalRowsAffected, nil
 }
 
 func (s *SQLStore) duplicateBlock(db sq.BaseRunner, boardID string, blockID string, userID string, asTemplate bool) ([]model.Block, error) {
