+++ server/services/store/sqlstore/params.go
@@ -18,7 +18,6 @@ type Params struct {
 	IsPlugin         bool
 	NewMutexFn       MutexFactory
 	PluginAPI        *plugin.API
-	SkipTemplateInit bool
 }
 
 func (p Params) CheckValid() error {
