+++ server/services/store/sqlstore/public_methods.go
@@ -648,30 +648,6 @@ func (s *SQLStore) RemoveDefaultTemplates(boards []*model.Board) error {
 
 }
 
-func (s *SQLStore) RunDataRetention(globalRetentionDate int64, batchSize int64) (int64, error) {
-	if s.dbType == model.SqliteDBType {
-		return s.runDataRetention(s.db, globalRetentionDate, batchSize)
-	}
-	tx, txErr := s.db.BeginTx(context.Background(), nil)
-	if txErr != nil {
-		return 0, txErr
-	}
-	result, err := s.runDataRetention(tx, globalRetentionDate, batchSize)
-	if err != nil {
-		if rollbackErr := tx.Rollback(); rollbackErr != nil {
-			s.logger.Error("transaction rollback error", mlog.Err(rollbackErr), mlog.String("methodName", "RunDataRetention"))
-		}
-		return 0, err
-	}
-
-	if err := tx.Commit(); err != nil {
-		return 0, err
-	}
-
-	return result, nil
-
-}
-
 func (s *SQLStore) SaveMember(bm *model.BoardMember) (*model.BoardMember, error) {
 	return s.saveMember(s.db, bm)
