+++ server/services/store/store.go
@@ -134,9 +134,6 @@ type Store interface {
 	RemoveDefaultTemplates(boards []*model.Board) error
 	GetTemplateBoards(teamID, userID string) ([]*model.Board, error)
 
-	// @withTransaction
-	RunDataRetention(globalRetentionDate int64, batchSize int64) (int64, error)
-
 	DBType() string
 
 	IsErrNotFound(err error) bool
